ntfsprogs/mkntfs: improve error when creating NTFS on a plain file
mkntfs is special among all other mkfs utilities in that it'd refuse
to create an FS over a plain file (which is often needed for testing
purposes). This might seem like an odd choice, but it has to do with
how NTFS works, and the error is there to make sure things wouldn't go
south for a user.

However, it lacked the explanation what exactly may be wrong, not to
mention it also didn't mention that FS could be created by passing
`--force`. Fix that.

Fixes: https://github.com/tuxera/ntfs-3g/issues/145